% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veWaningData.R
\name{veWaningData}
\alias{veWaningData}
\title{Toy Dataset For Illustration}
\format{
veWaningData is a data.frame containing 30,000 participants The
  data.frame contains 8 columns, 
  \describe{
  \item{X1}{A dichotomous baseline covariate; generated from B(0.5).}
  \item{X2}{A continuous baseline covariates; generated from N(45,10^2).}
  \item{E}{The entry time into the study.}
  \item{A}{The indicator of vaccination at entry
           (1 = vaccinated at entry; 0 = not vaccinated)}
  \item{U}{The time to infection; NA if participant did not become infected.}
  \item{R}{Time to unblinding.}
  \item{Gam}{The dynamic under which unblinding occurred; 0: infection
    occurred prior to unblinding; 1: participant requested unblinding
    prior to study defined unblinding period; 2: participant was
    unblinded during the study defined unblinding period.}
  \item{Psi}{Indicator of whether subject received vaccination at time of
    unblinding (1: received vaccination at time R, 0 otherwise).}
  }
}
\usage{
data(veWaningData)
}
\description{
This data set is provided for the purposes of illustrating the use of
the software.
}
\keyword{datasets}
